+++ packages/containers-shared/index.ts
@@ -16,3 +16,5 @@ export * from "./src/registry";
 export * from "./src/images";
 export * from "./src/spinner";
 export * from "./src/dev-options";
+export * from "./src/ssh";
+export * from "./src/registry-images";