+++ packages/wrangler/src/pages/run-workers-deploy.ts
@@ -5,6 +5,7 @@ import {
 	fetchPagedListResult,
 	fetchResult,
 } from "../cfetch";
+import { createCloudflareClient } from "../cfetch/internal";
 import { readConfig } from "../config";
 import { runDeployCommandHandler, type DeployArgs } from "../deploy";
 import { confirm, prompt, select } from "../dialogs";
@@ -49,6 +50,9 @@ export async function runPagesToWorkersDeploy(
 		profile: undefined,
 		triggers: undefined,
 		routes: undefined,
+		zone: undefined,
+		zoneId: undefined,
+		experimentalRouteZones: false,
 		domains: undefined,
 		metafile: undefined,
 		logpush: undefined,
@@ -97,6 +101,7 @@ export async function runPagesToWorkersDeploy(
 		await run(experimentalFlags, async () => {
 			initDeployHelpersContext({
 				logger,
+				createCloudflareClient,
 				fetchResult,
 				fetchListResult,
 				fetchPagedListResult,